SCB-909 Optimize package of load balance.
