+++ omega/omega-spring-starter/src/main/java/org/apache/servicecomb/saga/omega/spring/OmegaSpringConfig.java
@@ -19,12 +19,12 @@ package org.apache.servicecomb.saga.omega.spring;
 
 import com.google.common.collect.ImmutableList;
 import org.apache.servicecomb.saga.omega.connector.grpc.AlphaClusterConfig;
-import org.apache.servicecomb.saga.omega.connector.grpc.FastestSender;
+import org.apache.servicecomb.saga.omega.connector.grpc.core.FastestSender;
+import org.apache.servicecomb.saga.omega.connector.grpc.core.LoadBalanceContext;
+import org.apache.servicecomb.saga.omega.connector.grpc.core.LoadBalanceContextBuilder;
+import org.apache.servicecomb.saga.omega.connector.grpc.core.TransactionType;
 import org.apache.servicecomb.saga.omega.connector.grpc.saga.SagaLoadBalanceSender;
-import org.apache.servicecomb.saga.omega.connector.grpc.tcc.LoadBalanceContext;
-import org.apache.servicecomb.saga.omega.connector.grpc.tcc.LoadBalanceContextBuilder;
 import org.apache.servicecomb.saga.omega.connector.grpc.tcc.TccLoadBalanceSender;
-import org.apache.servicecomb.saga.omega.connector.grpc.tcc.TransactionType;
 import org.apache.servicecomb.saga.omega.context.CallbackContext;
 import org.apache.servicecomb.saga.omega.context.IdGenerator;
 import org.apache.servicecomb.saga.omega.context.OmegaContext;
@@ -116,29 +116,6 @@ class OmegaSpringConfig {
         reconnectDelay).build();
     return loadBalanceSenderContext;
   }
-
-//  @Bean
-//  SagaMessageSender grpcMessageSender(
-//      AlphaClusterConfig alphaClusterConfig,
-//      ServiceConfig serviceConfig,
-//      @Value("${omega.connection.reconnectDelay:3000}") int reconnectDelay) {
-//
-//    final SagaMessageSender sender = new LoadBalancedClusterMessageSender(
-//        alphaClusterConfig,
-//        serviceConfig,
-//        reconnectDelay);
-//
-//    sender.onConnected();
-//
-//    Runtime.getRuntime().addShutdownHook(new Thread(new Runnable() {
-//      @Override
-//      public void run() {
-//        sender.onDisconnected();
-//        sender.close();
-//      }
-//    }));
-//    return sender;
-//  }
 
   @Bean
   SagaMessageSender sagaLoadBalanceSender(@Qualifier("sagaLoadContext") LoadBalanceContext loadBalanceSenderContext) {
